requirements.txt
python-dotenv
openai>=1.0.0
icalendar
pytz
